few fixes
