+++ README.md
@@ -51,11 +51,12 @@ checkout [examples]('./examples/dothttpazure.http')
 - [x] properties file
 - [x] support various data types like json, text (headers should be added automatically)
 - [ ] update README.md to explain syntax and examples to refer
-- [ ] tests
+- [x] tests
+- [ ] integration tests
 - [ ] file uploads
-- [ ] handle 200, 300, 400, 500 exception (user readable messages)
+- [x] handle 200, 300, 400, 500 exception (user readable messages)
 - [ ] output should have rolling output format
-- [ ] multiple logging options
+- [ ] multiple logging options ?
 - [x] should handle exceptions, only message should be shown.
 - [x] debug mode to show logs
 - [ ] multi form upload
